+++ tests/test_optics.py
@@ -329,14 +329,14 @@ def test_OpticalPSF_flux_scaling():
         defocus=test_defocus, astig1=test_astig1, astig2=test_astig2, flux=test_flux)
     obj *= 2.
     np.testing.assert_almost_equal(
-        obj.getFlux(), test_flux * 2., decimal=param_decimal,
+        obj.flux, test_flux * 2., decimal=param_decimal,
         err_msg="Flux param inconsistent after __imul__.")
     obj = galsim.OpticalPSF(
         lam_over_diam=test_loD, oversampling=test_oversampling, pad_factor=test_pad_factor,
         defocus=test_defocus, astig1=test_astig1, astig2=test_astig2, flux=test_flux)
     obj /= 2.
     np.testing.assert_almost_equal(
-        obj.getFlux(), test_flux / 2., decimal=param_decimal,
+        obj.flux, test_flux / 2., decimal=param_decimal,
         err_msg="Flux param inconsistent after __idiv__.")
     obj = galsim.OpticalPSF(
         lam_over_diam=test_loD, oversampling=test_oversampling, pad_factor=test_pad_factor,
@@ -344,11 +344,11 @@ def test_OpticalPSF_flux_scaling():
     obj2 = obj * 2.
     # First test that original obj is unharmed...
     np.testing.assert_almost_equal(
-        obj.getFlux(), test_flux, decimal=param_decimal,
+        obj.flux, test_flux, decimal=param_decimal,
         err_msg="Flux param inconsistent after __rmul__ (original).")
     # Then test new obj2 flux
     np.testing.assert_almost_equal(
-        obj2.getFlux(), test_flux * 2., decimal=param_decimal,
+        obj2.flux, test_flux * 2., decimal=param_decimal,
         err_msg="Flux param inconsistent after __rmul__ (result).")
     obj = galsim.OpticalPSF(
         lam_over_diam=test_loD, oversampling=test_oversampling, pad_factor=test_pad_factor,
@@ -356,11 +356,11 @@ def test_OpticalPSF_flux_scaling():
     obj2 = 2. * obj
     # First test that original obj is unharmed...
     np.testing.assert_almost_equal(
-        obj.getFlux(), test_flux, decimal=param_decimal,
+        obj.flux, test_flux, decimal=param_decimal,
         err_msg="Flux param inconsistent after __mul__ (original).")
     # Then test new obj2 flux
     np.testing.assert_almost_equal(
-        obj2.getFlux(), test_flux * 2., decimal=param_decimal,
+        obj2.flux, test_flux * 2., decimal=param_decimal,
         err_msg="Flux param inconsistent after __mul__ (result).")
     obj = galsim.OpticalPSF(
         lam_over_diam=test_loD, oversampling=test_oversampling, pad_factor=test_pad_factor,
@@ -368,11 +368,11 @@ def test_OpticalPSF_flux_scaling():
     obj2 = obj / 2.
     # First test that original obj is unharmed...
     np.testing.assert_almost_equal(
-        obj.getFlux(), test_flux, decimal=param_decimal,
+        obj.flux, test_flux, decimal=param_decimal,
         err_msg="Flux param inconsistent after __div__ (original).")
     # Then test new obj2 flux
     np.testing.assert_almost_equal(
-        obj2.getFlux(), test_flux / 2., decimal=param_decimal,
+        obj2.flux, test_flux / 2., decimal=param_decimal,
         err_msg="Flux param inconsistent after __div__ (result).")